refactor(svelte): split plot-shared-services into announcer and semantic-keys
