--- packages/svelte/tests/plot-shared-services.test.ts
@@ -1,216 +1,0 @@
-/**
- * Unit tests for plot shared services (announcer + semantic keys).
- * Factories own effects — instantiate under `$effect.root` and destroy.
- */
-import { flushSync } from "svelte";
-import { describe, expect, it } from "vitest";
-
-import { runPipeline, type RenderModel } from "@ggsvelte/core";
-import { aes, gg } from "@ggsvelte/spec";
-
-import { createSourceIdentityTracker } from "../src/lib/plot-semantic-keys.js";
-import {
-  createPlotAnnouncer,
-  createSemanticKeyService,
-} from "../src/lib/plot-shared-services.svelte.js";
-import { withEffectRoot, withFlushedEffectRoot } from "./helpers/effect-root.svelte.js";
-import { reactiveBox } from "./helpers/reactive-box.svelte.js";
-
-const rows = [
-  { id: "a", x: 1, y: 10 },
-  { id: "b", x: 2, y: 20 },
-];
-
-function minimalModel(): RenderModel {
-  return runPipeline(
-    gg(rows, aes({ x: "x", y: "y", color: "id" }))
-      .geomPoint()
-      .spec(),
-    {
-      width: 400,
-      height: 300,
-    },
-  );
-}
-
-describe("createPlotAnnouncer", () => {
-  it("announce clears then sets after a microtask (live-region re-announce)", async () => {
-    const announcer = createPlotAnnouncer();
-    expect(announcer.text).toBe("");
-    announcer.announce("hello");
-    expect(announcer.text).toBe("");
-    await Promise.resolve();
-    expect(announcer.text).toBe("hello");
-    announcer.announce("hello");
-    expect(announcer.text).toBe("");
-    await Promise.resolve();
-    expect(announcer.text).toBe("hello");
-  });
-
-  it("clear is synchronous and never swallows a message queued in the same tick", async () => {
-    const announcer = createPlotAnnouncer();
-    // announce then clear then announce: only the last message may win, and a
-    // clear must not queue a blank that lands after a later announce.
-    announcer.announce("first");
-    announcer.clear();
-    expect(announcer.text).toBe("");
-    announcer.announce("second");
-    await Promise.resolve();
-    expect(announcer.text).toBe("second");
-  });
-});
-
-describe("createSemanticKeyService", () => {
-  it("keeps stable identity across a data-preserving re-render and flags key mutation as unstable", () => {
-    // Mutable rows: same array identity (same data token) across renders.
-    const mutableRows = [
-      { id: "a", x: 1, y: 10 },
-      { id: "b", x: 2, y: 20 },
-    ];
-    const buildModel = () =>
-      runPipeline(
-        gg(mutableRows, aes({ x: "x", y: "y", color: "id" }))
-          .geomPoint()
-          .spec(),
-        { width: 400, height: 300 },
-      );
-    const model = reactiveBox(buildModel());
-    const tracker = createSourceIdentityTracker();
-    const diagnostics: string[] = [];
-
-    const { value: service, destroy } = withFlushedEffectRoot(() => {
-      const created = createSemanticKeyService({
-        model: () => model.value,
-        assembled: () => ({
-          data: mutableRows,
-          layers: [{ geom: "point" }],
-          aes: { x: { field: "x" }, y: { field: "y" }, color: { field: "id" } },
-        }),
-        datumKey: () => "id",
-        data: () => mutableRows,
-        spec: () => null,
-        sourceIdentity: (value) => tracker.sourceIdentity(value),
-        deliverDiagnostic: (d) => {
-          diagnostics.push(d.code);
-        },
-      });
-      created.registerEffects();
-      return created;
-    });
-
-    const first = [service.keyAt(0), service.keyAt(1)];
-    expect(first).toEqual(["a", "b"]);
-    expect(diagnostics).toEqual([]);
-
-    // Data-preserving re-render: NEW model, same data token → same keys, no
-    // diagnostics. (model is reactive state, so the swap re-resolves keys.)
-    const previous = model.value;
-    model.set(buildModel());
-    flushSync();
-    expect([service.keyAt(0), service.keyAt(1)]).toEqual(first);
-    expect(diagnostics).toEqual([]);
-    previous.dispose();
-
-    // Key mutation under the SAME data token: priorKeys must catch it — the
-    // mutated row's key becomes null and an unstable-key diagnostic fires.
-    // This also proves the re-render above genuinely re-resolved (the same
-    // code path now observably reacts to the model swap).
-    mutableRows[0].id = "z";
-    const beforeMutation = model.value;
-    model.set(buildModel());
-    flushSync();
-    expect(service.keyAt(0)).toBeNull();
-    expect(service.keyAt(1)).toBe("b");
-    expect(diagnostics).toContain("INTERACTION_UNSTABLE_KEY");
-    beforeMutation.dispose();
-    model.value.dispose();
-    destroy();
-  });
-
-  it("delivers diagnostics once per change (behavior, not implementation)", () => {
-    const duplicateRows = () => [
-      { id: "a", x: 1, y: 1 },
-      { id: "a", x: 2, y: 2 },
-    ];
-    const buildModel = (data: { id: string; x: number; y: number }[]) =>
-      runPipeline(
-        gg(data, aes({ x: "x", y: "y" }))
-          .geomPoint()
-          .spec(),
-        {
-          width: 200,
-          height: 200,
-        },
-      );
-    const dataBox = reactiveBox(duplicateRows());
-    const modelBox = reactiveBox<RenderModel | null>(buildModel(dataBox.value));
-    const diagnostics: string[] = [];
-    const tracker = createSourceIdentityTracker();
-
-    const { destroy } = withFlushedEffectRoot(() => {
-      const created = createSemanticKeyService({
-        model: () => modelBox.value,
-        assembled: () => ({
-          data: dataBox.value,
-          layers: [{ geom: "point" }],
-          aes: { x: { field: "x" }, y: { field: "y" } },
-        }),
-        datumKey: () => "id",
-        data: () => dataBox.value,
-        spec: () => null,
-        sourceIdentity: (value) => tracker.sourceIdentity(value),
-        deliverDiagnostic: (d) => {
-          diagnostics.push(d.code);
-        },
-      });
-      created.registerEffects();
-      return created;
-    });
-
-    // Duplicate "a"/"a" keys → exactly one duplicate diagnostic on first resolve.
-    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY"]);
-    // Re-flush without changing any dependency must not re-deliver.
-    flushSync();
-    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY"]);
-
-    // A REAL change (new data identity, same duplicate condition) re-resolves
-    // and delivers the diagnostic exactly once more — once per change, not
-    // once per flush and not zero.
-    const previous = modelBox.value;
-    dataBox.set(duplicateRows());
-    modelBox.set(buildModel(dataBox.value));
-    flushSync();
-    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY", "INTERACTION_DUPLICATE_KEY"]);
-    previous?.dispose();
-    modelBox.value?.dispose();
-    destroy();
-  });
-
-  it("does not invoke deps during construction", () => {
-    let calls = 0;
-    const counting = <T>(value: T) => {
-      calls++;
-      return value;
-    };
-    const model = minimalModel();
-    const { destroy } = withEffectRoot(() =>
-      createSemanticKeyService({
-        model: () => counting(model),
-        assembled: () => counting(null),
-        datumKey: () => counting("id" as const),
-        data: () => counting(rows),
-        spec: () => counting(null),
-        sourceIdentity: (v) => {
-          calls++;
-          return String(v);
-        },
-        deliverDiagnostic: () => {
-          calls++;
-        },
-      }),
-    );
-    expect(calls).toBe(0);
-    destroy();
-    model.dispose();
-  });
-});
