+++ packages/svelte/tests/runtime/semantic-keys.test.ts
@@ -1,16 +1,43 @@
+/**
+ * Unit tests for semantic-key pure helpers + reactive service.
+ * Factories own effects — instantiate under `$effect.root` and destroy.
+ */
+import { flushSync } from "svelte";
 import { describe, expect, it } from "vitest";
 
 import type { CellValue } from "@ggsvelte/core";
+import { runPipeline, type RenderModel } from "@ggsvelte/core";
+import { aes, gg } from "@ggsvelte/spec";
 
-import { INTERACTION_DIAGNOSTIC_CATALOG } from "../src/lib/interaction.js";
+import { INTERACTION_DIAGNOSTIC_CATALOG } from "../../src/lib/interaction.js";
 import {
   createSourceIdentityTracker,
+  createSemanticKeyService,
   dataIdentityEpochToken,
   resolveSemanticKeys,
   resolveSemanticKeysForPlot,
   type SemanticKeyCandidate,
   type SemanticKeyModelView,
-} from "../src/lib/plot-semantic-keys.js";
+} from "../../src/lib/runtime/semantic-keys.svelte.js";
+import { withEffectRoot, withFlushedEffectRoot } from "../helpers/effect-root.svelte.js";
+import { reactiveBox } from "../helpers/reactive-box.svelte.js";
+
+const rows = [
+  { id: "a", x: 1, y: 10 },
+  { id: "b", x: 2, y: 20 },
+];
+
+function minimalModel(): RenderModel {
+  return runPipeline(
+    gg(rows, aes({ x: "x", y: "y", color: "id" }))
+      .geomPoint()
+      .spec(),
+    {
+      width: 400,
+      height: 300,
+    },
+  );
+}
 
 function modelView(options: {
   candidates?: SemanticKeyCandidate[];
@@ -22,7 +49,7 @@ function modelView(options: {
   }>;
 }): SemanticKeyModelView {
   const candidates = options.candidates ?? [];
-  const rows = options.rows ?? [];
+  const modelRows = options.rows ?? [];
   const lineage = options.lineage ?? new Map<number, readonly number[]>();
   return {
     candidateCount: candidates.length,
@@ -33,11 +60,15 @@ function modelView(options: {
       return lineage.get(lineageId) ?? [];
     },
     row(rowIndex: number) {
-      return rows[rowIndex] ?? null;
+      return modelRows[rowIndex] ?? null;
     },
     layers: options.layers ?? [],
   };
 }
+
+// ---------------------------------------------------------------------------
+// Pure helpers (from plot-semantic-keys.test.ts)
+// ---------------------------------------------------------------------------
 
 describe("dataIdentityEpochToken", () => {
   it("returns no-data when assembled is null", () => {
@@ -355,5 +386,237 @@ describe("resolveSemanticKeys", () => {
     expect(result.diagnostics).toEqual([]);
     expect(priorKeys.get("dataB:0")).toBe("b");
     expect(priorKeys.get("dataA:0")).toBe("a");
+  });
+});
+
+// ---------------------------------------------------------------------------
+// Reactive service (from plot-shared-services.test.ts)
+// ---------------------------------------------------------------------------
+
+describe("createSemanticKeyService", () => {
+  it("keeps stable identity across a data-preserving re-render and flags key mutation as unstable", () => {
+    // Mutable rows: same array identity (same data token) across renders.
+    const mutableRows = [
+      { id: "a", x: 1, y: 10 },
+      { id: "b", x: 2, y: 20 },
+    ];
+    const buildModel = () =>
+      runPipeline(
+        gg(mutableRows, aes({ x: "x", y: "y", color: "id" }))
+          .geomPoint()
+          .spec(),
+        { width: 400, height: 300 },
+      );
+    const model = reactiveBox(buildModel());
+    const tracker = createSourceIdentityTracker();
+    const diagnostics: string[] = [];
+
+    const { value: service, destroy } = withFlushedEffectRoot(() => {
+      const created = createSemanticKeyService({
+        model: () => model.value,
+        assembled: () => ({
+          data: mutableRows,
+          layers: [{ geom: "point" }],
+          aes: { x: { field: "x" }, y: { field: "y" }, color: { field: "id" } },
+        }),
+        datumKey: () => "id",
+        data: () => mutableRows,
+        spec: () => null,
+        sourceIdentity: (value) => tracker.sourceIdentity(value),
+        deliverDiagnostic: (d) => {
+          diagnostics.push(d.code);
+        },
+      });
+      created.registerEffects();
+      return created;
+    });
+
+    const first = [service.keyAt(0), service.keyAt(1)];
+    expect(first).toEqual(["a", "b"]);
+    expect(diagnostics).toEqual([]);
+
+    // Data-preserving re-render: NEW model, same data token → same keys, no
+    // diagnostics. (model is reactive state, so the swap re-resolves keys.)
+    const previous = model.value;
+    model.set(buildModel());
+    flushSync();
+    expect([service.keyAt(0), service.keyAt(1)]).toEqual(first);
+    expect(diagnostics).toEqual([]);
+    previous.dispose();
+
+    // Key mutation under the SAME data token: priorKeys must catch it — the
+    // mutated row's key becomes null and an unstable-key diagnostic fires.
+    // This also proves the re-render above genuinely re-resolved (the same
+    // code path now observably reacts to the model swap).
+    mutableRows[0].id = "z";
+    const beforeMutation = model.value;
+    model.set(buildModel());
+    flushSync();
+    expect(service.keyAt(0)).toBeNull();
+    expect(service.keyAt(1)).toBe("b");
+    expect(diagnostics).toContain("INTERACTION_UNSTABLE_KEY");
+    beforeMutation.dispose();
+    model.value.dispose();
+    destroy();
+  });
+
+  it("delivers diagnostics once per change (behavior, not implementation)", () => {
+    const duplicateRows = () => [
+      { id: "a", x: 1, y: 1 },
+      { id: "a", x: 2, y: 2 },
+    ];
+    const buildModel = (data: { id: string; x: number; y: number }[]) =>
+      runPipeline(
+        gg(data, aes({ x: "x", y: "y" }))
+          .geomPoint()
+          .spec(),
+        {
+          width: 200,
+          height: 200,
+        },
+      );
+    const dataBox = reactiveBox(duplicateRows());
+    const modelBox = reactiveBox<RenderModel | null>(buildModel(dataBox.value));
+    const diagnostics: string[] = [];
+    const tracker = createSourceIdentityTracker();
+
+    const { destroy } = withFlushedEffectRoot(() => {
+      const created = createSemanticKeyService({
+        model: () => modelBox.value,
+        assembled: () => ({
+          data: dataBox.value,
+          layers: [{ geom: "point" }],
+          aes: { x: { field: "x" }, y: { field: "y" } },
+        }),
+        datumKey: () => "id",
+        data: () => dataBox.value,
+        spec: () => null,
+        sourceIdentity: (value) => tracker.sourceIdentity(value),
+        deliverDiagnostic: (d) => {
+          diagnostics.push(d.code);
+        },
+      });
+      created.registerEffects();
+      return created;
+    });
+
+    // Duplicate "a"/"a" keys → exactly one duplicate diagnostic on first resolve.
+    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY"]);
+    // Re-flush without changing any dependency must not re-deliver.
+    flushSync();
+    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY"]);
+
+    // A REAL change (new data identity, same duplicate condition) re-resolves
+    // and delivers the diagnostic exactly once more — once per change, not
+    // once per flush and not zero.
+    const previous = modelBox.value;
+    dataBox.set(duplicateRows());
+    modelBox.set(buildModel(dataBox.value));
+    flushSync();
+    expect(diagnostics).toEqual(["INTERACTION_DUPLICATE_KEY", "INTERACTION_DUPLICATE_KEY"]);
+    previous?.dispose();
+    modelBox.value?.dispose();
+    destroy();
+  });
+
+  it("does not invoke deps during construction", () => {
+    let calls = 0;
+    const counting = <T>(value: T) => {
+      calls++;
+      return value;
+    };
+    const model = minimalModel();
+    const { destroy } = withEffectRoot(() =>
+      createSemanticKeyService({
+        model: () => counting(model),
+        assembled: () => counting(null),
+        datumKey: () => counting("id" as const),
+        data: () => counting(rows),
+        spec: () => counting(null),
+        sourceIdentity: (v) => {
+          calls++;
+          return String(v);
+        },
+        deliverDiagnostic: () => {
+          calls++;
+        },
+      }),
+    );
+    expect(calls).toBe(0);
+    destroy();
+    model.dispose();
+  });
+
+  it("legendEntryKeyIndex getter and keysForLegend both re-resolve on model change", () => {
+    // ACCESS CONTRACT: legendEntryKeyIndex is a reactive getter (not a plain
+    // property snapshot). Mutating the model dep must refresh BOTH the getter
+    // and keysForLegend — a snapped Map would leave both paths stale.
+    const buildModel = (data: { id: string; x: number; y: number }[]) =>
+      runPipeline(
+        gg(data, aes({ x: "x", y: "y", color: "id" }))
+          .geomPoint()
+          .spec(),
+        { width: 400, height: 300 },
+      );
+    const dataA = [
+      { id: "a", x: 1, y: 10 },
+      { id: "b", x: 2, y: 20 },
+    ];
+    const dataB = [
+      { id: "c", x: 3, y: 30 },
+      { id: "d", x: 4, y: 40 },
+    ];
+    const dataBox = reactiveBox(dataA);
+    const modelBox = reactiveBox(buildModel(dataBox.value));
+    const tracker = createSourceIdentityTracker();
+
+    const { value: service, destroy } = withFlushedEffectRoot(() => {
+      const created = createSemanticKeyService({
+        model: () => modelBox.value,
+        assembled: () => ({
+          data: dataBox.value,
+          layers: [{ geom: "point" }],
+          aes: { x: { field: "x" }, y: { field: "y" }, color: { field: "id" } },
+        }),
+        datumKey: () => "id",
+        data: () => dataBox.value,
+        spec: () => null,
+        sourceIdentity: (value) => tracker.sourceIdentity(value),
+        deliverDiagnostic: () => {},
+      });
+      created.registerEffects();
+      return created;
+    });
+
+    const action = {
+      identity: { scale: "color", entryIndex: 0 },
+      entry: { value: "a", label: "a" },
+      source: "keyboard" as const,
+    };
+    // Discrete color legend: entry 0 maps to first category's semantic key.
+    const indexBefore = service.legendEntryKeyIndex;
+    expect(indexBefore.size).toBeGreaterThan(0);
+    const keysBefore = service.keysForLegend(action);
+    expect(keysBefore).toEqual(["a"]);
+    // Snapshot identity of the Map itself — a plain property would pin this.
+    const mapRefBefore = indexBefore;
+
+    const previous = modelBox.value;
+    dataBox.set(dataB);
+    modelBox.set(buildModel(dataBox.value));
+    flushSync();
+
+    const indexAfter = service.legendEntryKeyIndex;
+    const keysAfter = service.keysForLegend(action);
+    // Getter must re-read the derived Map (not return a construction-time snap).
+    expect(indexAfter).not.toBe(mapRefBefore);
+    // Keys for entry 0 must follow the new model categories (c, not a).
+    expect(keysAfter).toEqual(["c"]);
+    expect(service.keyAt(0)).toBe("c");
+    expect(service.keyAt(1)).toBe("d");
+
+    previous.dispose();
+    modelBox.value.dispose();
+    destroy();
   });
 });
